+++ tests/Unit/Like/Events/ModelWasUndislikedTest.php
@@ -9,28 +9,30 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Tests\Laravel\Likeable\Unit\Events;
+namespace Cog\Tests\Laravel\Likeable\Unit\Like\Events;
 
-use Cog\Laravel\Likeable\Like\Events\ModelWasLiked;
+use Cog\Laravel\Likeable\Like\Events\ModelWasUndisliked;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
 /**
- * Class ModelWasLikedTest.
+ * Class ModelWasUndislikedTest.
  *
- * @package Cog\Tests\Laravel\Likeable\Unit\Events
+ * @package Cog\Tests\Laravel\Likeable\Unit\Like\Events
  */
-class ModelWasLikedTest extends TestCase
+class ModelWasUndislikedTest extends TestCase
 {
     use DatabaseTransactions;
 
     /** @test */
     public function it_can_fire_model_was_liked_event()
     {
-        $this->expectsEvents(ModelWasLiked::class);
+        $this->expectsEvents(ModelWasUndisliked::class);
 
         $entity = factory(Entity::class)->create();
-        $entity->like(1);
+        $entity->dislike(1);
+
+        $entity->undislike(1);
     }
 }
