+++ tests/Unit/LikeCounter/Models/LikeCounterTest.php
@@ -9,15 +9,16 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Tests\Laravel\Likeable\Unit\Models;
+namespace Cog\Tests\Laravel\Likeable\Unit\LikeCounter\Models;
 
 use Cog\Laravel\Likeable\LikeCounter\Models\LikeCounter;
+use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 
 /**
  * Class LikeCounterTest.
  *
- * @package Cog\Tests\Laravel\Likeable\Unit\Models
+ * @package Cog\Tests\Laravel\Likeable\Unit\LikeCounter\Models
  */
 class LikeCounterTest extends TestCase
 {
@@ -49,5 +50,15 @@ class LikeCounterTest extends TestCase
         ]);
 
         $this->assertTrue(is_int($like->count));
+    }
+
+    /** @test */
+    public function it_can_belong_to_likeable_model()
+    {
+        $entity = factory(Entity::class)->create();
+
+        $entity->like(1);
+
+        $this->assertInstanceOf(Entity::class, LikeCounter::first()->likeable);
     }
 }
